Refactor schema implementation logic in template.

See merge request igrp-3_0/spring-engine!17

+++ src/modules/model/generateModel.ts
@@ -1,16 +1,22 @@
 import { Attribute, ModelConfig, RenderContext } from '../../interfaces/types';
 import { renderTemplate } from '../common/renderTemplate';
-import { ERROR_MESSAGE, EXTENSIONS, TEMPLATES } from '../../utils/constants';
+import { DIRECTORIES, ERROR_MESSAGE, EXTENSIONS, PROJECT_STRUCTURE_STYLE, TEMPLATES } from '../../utils/constants';
 import { saveToFile } from '../common/saveToFile';
-import { getModelOutputDir } from '../../utils/helpers';
+import { getDDDModelOutputDir, getModelOutputDir } from '../../utils/helpers';
 import path from 'path';
 import fs from 'fs-extra';
 import { updatePermissions } from '../permission/permissionManagement';
 import { saveModelConfig } from './saveModelConfig';
 
 export const generateModel = async (context: RenderContext<ModelConfig>) => {
+  let modelOutputPath: string;
+
+  if(context.baseConfig.projectStructureStyle == PROJECT_STRUCTURE_STYLE.DOMAIN_DRIVEN_DESIGN)
+    modelOutputPath = getDDDModelOutputPath(context);
+  else
+    modelOutputPath = getModelOutputPath(context);
+
   const template = await renderModel(context);
-  const modelOutputPath = getModelOutputPath(context);
 
   const errosUniqueConstraints = validarUniqueConstraints(context.resourceConfig);
 
@@ -18,71 +24,83 @@ export const generateModel = async (context: RenderContext<ModelConfig>) => {
     throw new Error(`Erros de uniqueConstraints encontrados:\n${errosUniqueConstraints.join('\n')}`);
   }
 
-  // Before creating the model with the new configuration, the engine checks if a model directory already exists.
-  // If it does, the old model directory will be deleted to ensure a clean setup for the new model.
-  const modelDirector = modelDirectory(context);
-  if (await fs.pathExists(modelDirector)) await fs.rm(modelDirector, { recursive: true });
+  if (context.baseConfig.projectStructureStyle != PROJECT_STRUCTURE_STYLE.DOMAIN_DRIVEN_DESIGN) {
+    // Before creating the model with the new configuration, the engine checks if a model directory already exists.
+    // If it does, the old model directory will be deleted to ensure a clean setup for the new model.
+    const modelDirector = modelDirectory(context);
+    if (await fs.pathExists(modelDirector)) await fs.rm(modelDirector, { recursive: true });
+  }
 
   const {primaryKey} = context.resourceConfig
 
   if (primaryKey) {
-    const primaryKeyTemplate = await renderPrimaryKey(context);
     const primaryKeyPath = getPrimaryKeyModelOutputPath(context);
+    const primaryKeyTemplate = await renderPrimaryKey(context);
     await saveToFile(primaryKeyTemplate, primaryKeyPath);
-  } 
+  }
+
+  await saveToFile(template, modelOutputPath, true, DIRECTORIES.MODELS, context.resourceConfig.id, context.resourceConfig.module, context.basePath);
 
   await saveModelConfig(context.resourceConfig, context.basePath);
-
-  await saveToFile(template, modelOutputPath);
 
   // Once the model has been generated, we will assign the necessary permissions to its endpoints.
   // This ensures that the newly created model has the correct access rights configured 
   // for each endpoint based on its defined permissions.
-  await updatePermissions(context.basePath, context.resourceConfig.type);
+  await updatePermissions(context.resourceConfig.module ?? DIRECTORIES.SHARED, context.basePath, context.resourceConfig.type);
 
 };
 
 /**
  * Generates the model in the API using the provided configuration.
- * @param ontext - The configuration of the model including the model name and attributes.
+ * @param context - The configuration of the model including the model name and attributes.
  * @returns - A string representing the model generated from the template.
  * @throws - Throws an error if the model configuration is invalid or has no attributes.
  */
 const renderModel = async (context: RenderContext<ModelConfig>) => {
-  context.sqlAttributes = sqlUniquesAttributes(context.resourceConfig.attributes)
+
+  context.dateTimeAttributes = dateTimeUniqueAttributes(context.resourceConfig.attributes)
   context.mathAttributes = mathUniquesAttributes(context.resourceConfig.attributes)
 
   if (context.resourceConfig.attributes.length === 0) {
     throw ERROR_MESSAGE.EMPTY_ATTRIBUTE;
   }
 
-    // Validação e renderização de cada atributo com seu valor padrão
-    context.resourceConfig.attributes.forEach(attribute => {
-      if (attribute.defaultValue) {
-        const columnAnnotation = renderColumnWithDefault(attribute);
-      }
-    });
+  // Validação e renderização de cada atributo com seu valor padrão
+  context.resourceConfig.attributes.forEach(attribute => {
+    if (attribute.defaultValue) {
+      const columnAnnotation = renderColumnWithDefault(attribute);
+    }
+  });
 
     // Gerar as restrições únicas compostas
   context.uniqueConstraints = context.resourceConfig.uniqueConstraints || [];
 
   return await renderTemplate(TEMPLATES.DOMAIN_MODEL, context);
+
 };
 
-const renderPrimaryKey = async(context: RenderContext<ModelConfig>) => {
+const renderPrimaryKey = async (context: RenderContext<ModelConfig>) => {
   return await renderTemplate(TEMPLATES.DOMAIN_MODEL_PRIMARY_KEY, context);
-}
+};
 
-const sqlUniquesAttributes = (attributes: Attribute[]) => {
-  let sqlAttributes: string[] = [];
-  attributes.forEach(attribute => {
-    if (attribute.type === "Date" || attribute.type === "Time" || attribute.type === "Timestamp"){
-      sqlAttributes.push(attribute.type)
+const dateTimeUniqueAttributes = (attributes: Attribute[]) => {
+  let dateTimeAttributes: string[] = [];
+  attributes.forEach((attribute) => {
+    if (
+      attribute.type === 'LocalTime' ||
+      attribute.type === 'LocalDate' ||
+      attribute.type === 'LocalDateTime' ||
+      attribute.type === 'ZoneDateTime' ||
+      attribute.type === 'OffsetDateTime' ||
+      attribute.type === 'Instant'
+    ) {
+      dateTimeAttributes.push(attribute.type);
     }
-  })
+  });
 
-  return [...new Set(sqlAttributes)]
-}
+  return [...new Set(dateTimeAttributes)];
+};
+
 // Verefica existencia de atributos de alta precisão
 const mathUniquesAttributes = (attributes: Attribute[]) => {
   let mathAttributes: string[] = [];
@@ -99,7 +117,7 @@ const validateColumnDefault = (attribute: Attribute) => {
   const { type, defaultValue } = attribute;
 
   if (defaultValue) { // Só valida se defaultValue estiver presente
-    if (['Integer', 'BigInteger', 'BigDecimal', 'Long', 'Double', 'Float', 'Short', 'Byte'].includes(type)) {
+    if (['integer', 'biginteger', 'bigdecimal', 'long', 'double', 'float', 'short', 'byte'].includes(type)) {
       // Verifica se o valor padrão é um número válido
       if (isNaN(Number(defaultValue))) {
         throw new Error(`The default value "${defaultValue}" is not valid for the numeric type ${type}.`);
@@ -115,7 +133,7 @@ const validateColumnDefault = (attribute: Attribute) => {
         throw new Error(`The default value "${defaultValue}" is not valid for the boolean type ${type}.`);
       }
     } else {
-      console.warn(`No specific validation for the type ${type}.`);
+      //console.warn(`No specific validation for the type ${type}.`);
     }
   }
 };
@@ -124,7 +142,7 @@ const renderColumnWithDefault = (attribute: Attribute) => {
   validateColumnDefault(attribute);
 
   if (attribute.defaultValue) {
-    return `@ColumnDefault(${attribute.defaultValue})`;
+    return `@ColumnDefault('${attribute.defaultValue}')`;
   }
 
   return ''; // Caso não tenha valor padrão, não retorna a anotação
@@ -150,15 +168,47 @@ function validarUniqueConstraints(modelConfig: ModelConfig): string[] {
 }
 
 // Caminho onde o arquivo é salvo
-const getModelOutputPath = (context: RenderContext<ModelConfig>) => 
-  path.join(getModelOutputDir(context), `${context.resourceConfig.name}${EXTENSIONS.JAVA}`)
+const getModelOutputPath = (context: RenderContext<ModelConfig>) => {
+  const outputDir = getModelOutputDir(context)
+  context.fullPath = outputDir
+  return path.join(outputDir, `${context.resourceConfig.name}${EXTENSIONS.JAVA}`);
+}
 
-const getPrimaryKeyModelOutputPath = (context: RenderContext<ModelConfig>) =>
-  path.join(getModelOutputDir(context),`${context.resourceConfig.name}PrimaryKey${EXTENSIONS.JAVA}`);
+// Caminho onde o arquivo é salvo
+const getDDDModelOutputPath = (context: RenderContext<ModelConfig>) => {
+  const outputDir = getDDDModelOutputDir(context)
+  context.fullPath = outputDir
+  return path.join(outputDir, `${context.resourceConfig.name}${EXTENSIONS.JAVA}`);
+}
 
-const modelDirectory = (context: RenderContext<ModelConfig>) => 
-  path.join(getModelOutputDir(context))
+const getPrimaryKeyModelOutputPath = (context: RenderContext<ModelConfig>) => {
+  if(context.baseConfig.projectStructureStyle === PROJECT_STRUCTURE_STYLE.DOMAIN_DRIVEN_DESIGN) {
+    const outputDir = getDDDModelOutputDir(context)
+    context.fullPath = outputDir
+    return path.join(
+      outputDir,
+      `${context.resourceConfig.name}PrimaryKey${EXTENSIONS.JAVA}`,
+    );
+  }
+  else {
+    const outputDir = getModelOutputDir(context)
+    context.fullPath = outputDir
+    return path.join(
+      outputDir,
+      `${context.resourceConfig.name}PrimaryKey${EXTENSIONS.JAVA}`,
+    );
+  }
+}
 
-
-
-
+const modelDirectory = (context: RenderContext<ModelConfig>) => {
+  if(context.baseConfig.projectStructureStyle === PROJECT_STRUCTURE_STYLE.DOMAIN_DRIVEN_DESIGN) {
+    const outputDir = getDDDModelOutputDir(context)
+    context.fullPath = outputDir
+    return path.join(outputDir);
+  }
+  else {
+    const outputDir = getModelOutputDir(context)
+    context.fullPath = outputDir
+    return path.join(outputDir);
+  }
+}
